# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: nomad-mvp API
  version: 1.0.0
  description: |
    Core asynchronous workflows with SSE. Latency SLOs are user-perceived:
    ACK (HTTP accept), TTFU (first SSE event), Keep-alive (max gap ≤10s).
servers:
  - url: https://api.example.com
    description: Example server (replace in deployment)
...
paths:
...
  /auth/otp/start:
    post:
      tags: [Auth]
      summary: Start OTP flow (SMS)
      operationId: authOtpStart
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OtpStartRequest' }
      parameters:
        - $ref: '#/components/parameters/ExpectedUser'
        - $ref: '#/components/parameters/ExpectedSession'
        - $ref: '#/components/parameters/RequestOrigin'
      responses:
        '409': { $ref: '#/components/responses/Error409' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: OTP send status or captcha requirement
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpStartResponse' }
        '400': { $ref: '#/components/responses/Error400' }
        '403': { $ref: '#/components/responses/Error403' }
        '429':
          $ref: '#/components/responses/Error429'
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and create session
      operationId: authOtpVerify
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OtpVerifyRequest' }
      parameters:
        - $ref: '#/components/parameters/ExpectedUser'
        - $ref: '#/components/parameters/ExpectedSession'
        - $ref: '#/components/parameters/RequestOrigin'
      responses:
        '409': { $ref: '#/components/responses/Error409' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: Session created
          headers:
            X-Device-Id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpVerifyResponse'
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '429': { $ref: '#/components/responses/Error429' }
  /auth/native/otp/start:
    post:
      tags: [Auth]
      summary: Start OTP flow (SMS)
      operationId: native_authOtpStart
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OtpStartRequest' }
      parameters:
        - $ref: '#/components/parameters/ExpectedUser'
        - $ref: '#/components/parameters/ExpectedSession'
        - $ref: '#/components/parameters/NativeAudience'
        - $ref: '#/components/parameters/NativeBinding'
      responses:
        '409': { $ref: '#/components/responses/Error409' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: OTP send status or captcha requirement
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OtpStartResponse' }
        '400': { $ref: '#/components/responses/Error400' }
        '403': { $ref: '#/components/responses/Error403' }
        '429':
          $ref: '#/components/responses/Error429'
  /auth/native/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and create session
      operationId: native_authOtpVerify
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OtpVerifyRequest' }
      parameters:
        - $ref: '#/components/parameters/ExpectedUser'
        - $ref: '#/components/parameters/ExpectedSession'
        - $ref: '#/components/parameters/NativeAudience'
        - $ref: '#/components/parameters/NativeBinding'
      responses:
        '409': { $ref: '#/components/responses/Error409' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: Session created
          headers:
            X-Device-Id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NativeOtpVerifyResponse'
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '429': { $ref: '#/components/responses/Error429' }
  /me:
    get:
      tags: [Sessions]
      summary: Get current user and session
      operationId: me
      security:
        - SessionAuth: []
        - BearerAuth: []
      responses:
        '403': { $ref: '#/components/responses/Error403' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: Current user and session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401': { $ref: '#/components/responses/Error401' }
  /sessions/me:
    get:
      tags: [Sessions]
      summary: Get current session
      operationId: sessionsMe
      security:
        - SessionAuth: []
        - BearerAuth: []
      responses:
        '403': { $ref: '#/components/responses/Error403' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401': { $ref: '#/components/responses/Error401' }
  /sessions:
    get:
      tags: [Sessions]
      summary: List my sessions
      operationId: sessionsList
      security:
        - SessionAuth: []
        - BearerAuth: []
      responses:
        '403': { $ref: '#/components/responses/Error403' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: Sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401': { $ref: '#/components/responses/Error401' }
  /sessions/{id}:
    delete:
      tags: [Sessions]
      summary: Revoke a session
      operationId: sessionsDelete
      security:
        - SessionAuth: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestOrigin'
        - $ref: '#/components/parameters/ExpectedUser'
        - $ref: '#/components/parameters/ExpectedSession'
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '409': { $ref: '#/components/responses/Error409' }
        '400': { $ref: '#/components/responses/Error400' }
        '403': { $ref: '#/components/responses/Error403' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
  /ops/me:
    get:
      security: [{ SessionAuth: [] }]
      tags: [Auth]
      operationId: readOperatorAccess
      summary: Read currently granted desktop operator capabilities
      description: Web session only. Ordinary login, client roles and knowing this URL grant no operator authority. Grants are checked from the persistent authority on each request.
      parameters:
        - $ref: '#/components/parameters/ExpectedUser'
        - $ref: '#/components/parameters/ExpectedSession'
      responses:
        '200':
          description: Current effective scoped grants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorAccessResponse' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '503': { $ref: '#/components/responses/Error503' }
  /ops/access-check:
    post:
      security: [{ SessionAuth: [] }]
      tags: [Auth]
      operationId: checkOperatorAccess
      summary: Verify scoped write authority and create an audit receipt
      description: Locks the qualified caller and the exact current grant in the audit-write transaction. This writes only an access-check receipt, not business data or a permission grant. Retries retain operation_id; revoked or changed grants never replay a successful old receipt.
      parameters:
        - $ref: '#/components/parameters/RequestOrigin'
        - $ref: '#/components/parameters/ExpectedUser'
        - $ref: '#/components/parameters/ExpectedSession'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorAccessCheckRequest' }
      responses:
        '200':
          description: Durable authorization-check receipt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperatorAccessCheckResponse' }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '409': { $ref: '#/components/responses/Error409' }
        '503': { $ref: '#/components/responses/Error503' }
  /logout:
    post:
      tags: [Sessions]
      summary: Revoke the current session and clear its cookie
      operationId: logout
      description: Confirms durable current-session revocation before clearing the cookie; other device sessions remain valid. Unknown storage outcomes return 503 and must not be reported as successful logout. The same operation_id and expected public identity confirm a lost response even after cookie clearing; this receipt never authenticates a user or revokes a newer session.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      parameters:
        - $ref: '#/components/parameters/RequestOrigin'
        - $ref: '#/components/parameters/ExpectedUser'
        - $ref: '#/components/parameters/ExpectedSession'
      responses:
        '409': { $ref: '#/components/responses/Error409' }
        '403': { $ref: '#/components/responses/Error403' }
        '400': { $ref: '#/components/responses/Error400' }
        '503': { $ref: '#/components/responses/Error503' }
        '200':
          description: Logged out
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
...
components:
  parameters:
    RequestOrigin:
      in: header
      name: Origin
      required: true
      description: Exact configured browser origin required on unsafe requests; never derived from Host or forwarded headers.
      schema: { type: string }
    ExpectedUser:
      in: header
      name: X-Auth-User-Id
      required: false
      description: Required for authenticated browser resource requests. Login/OTP requests use anonymous when the initiating page is unauthenticated. Expected owner from the initiating page; only compared AFTER real session authentication, never a credential or an identity selector.
      schema: { type: string, maxLength: 128 }
    NativeAudience:
      in: header
      name: X-Nomad-Auth-Audience
      required: true
      description: Native fixed API origin. HTTPS required; browser Origin and Cookie are rejected.
      schema: { type: string, format: uri }
    NativeBinding:
      in: header
      name: X-Nomad-Login-Binding
      required: true
      description: Secret generated and retained by the native plugin. Never available to JavaScript or logs.
      schema: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' }
...
    ExpectedSession:
      in: header
      name: X-Auth-Session-Id
      required: false
      description: Required for authenticated browser resource requests. Login/OTP requests use anonymous when the initiating page is unauthenticated. Public session reference from the initiating page; stale or mismatched context returns AUTH_CONTEXT_CHANGED (409).
      schema: { type: string, maxLength: 128 }
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: __Host-nomad-sid-{public_session_id}
      description: HttpOnly Secure cookie family. Each login writes a distinct public-session suffix; delayed login/logout responses cannot overwrite newer credentials. The suffix is public and never authenticates without the opaque cookie value.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Native only. Requires HTTPS, X-Nomad-Auth-Audience matching AUTH_API_ORIGIN, and no Origin or Cookie. The credential is held by the native plugin and is never returned to JavaScript.
  schemas:
...
    PnvsCaptchaProof:
      type: object
      additionalProperties: false
      required: [lot_number, captcha_output, pass_token, gen_time]
      properties:
        lot_number: { type: string, minLength: 1, maxLength: 128 }
        captcha_output: { type: string, minLength: 1, maxLength: 4096 }
        pass_token: { type: string, minLength: 1, maxLength: 1024 }
        gen_time: { type: string, minLength: 1, maxLength: 32 }
    OperatorGrant:
      type: object
      required: [capability, scope, version]
      properties:
        capability: { type: string, enum: [places.correct, brand.rules, ops.workbench, ops.prompts, ops.rollout, ops.usage] }
        scope: { type: string, minLength: 1, maxLength: 128 }
        version: { type: integer, minimum: 1 }
    OperatorAccessResponse:
      type: object
      required: [user_id, grants]
      properties:
        user_id: { type: string, format: uuid }
        grants: { type: array, items: { $ref: '#/components/schemas/OperatorGrant' } }
    OperatorAccessCheckRequest:
      type: object
      additionalProperties: false
      required: [capability, scope, expected_grant_version, operation_id]
      properties:
        capability: { type: string, enum: [places.correct, brand.rules, ops.workbench, ops.prompts, ops.rollout, ops.usage] }
        scope: { type: string, minLength: 1, maxLength: 128 }
        expected_grant_version: { type: integer, minimum: 1 }
        operation_id: { type: string, format: uuid }
    OperatorAccessCheckResponse:
      type: object
      required: [receipt_id, verified_at]
      properties:
        receipt_id: { type: string, format: uuid }
        verified_at: { type: string, format: date-time }
...
    NativeOtpVerifyResponse:
      description: Internal HTTPS response consumed only by the native plugin. The plugin stores and strips the credential before resolving any JavaScript promise.
      allOf:
        - $ref: '#/components/schemas/OtpVerifyResponse'
        - type: object
          required: [native_session_credential]
          properties:
            native_session_credential: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' }
    OtpStartRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string, minLength: 6, maxLength: 32, pattern: '^\+?[0-9][0-9 -]{4,30}[0-9]$' }
        region: { type: string, minLength: 2, maxLength: 8, pattern: '^[A-Z]{2,8}$', default: CN }
        request_id: { type: string, format: uuid, description: Required in real mode; one browser intent retains this ID across an unknown-result retry. }
        captcha: { $ref: '#/components/schemas/PnvsCaptchaProof' }
        captcha_token: { type: string, minLength: 1, maxLength: 2048, nullable: true, description: Legacy isolated fixture only; never accepted as real PNVS proof. }
    OtpStartResponse:
      type: object
      required: [sent, retry_after_sec, captcha_required]
      properties:
        sent: { type: boolean, description: Provider accepted this send; does not prove handset delivery or login. }
        retry_after_sec: { type: integer, minimum: 1 }
        challenge_id: { type: string, format: uuid, nullable: true }
        delivery_state: { type: string, enum: [sent, unknown, captcha_required] }
        captcha_required: { type: boolean }
        captcha_provider: { type: string, nullable: true, enum: [tencent, aliyun-pnvs, fixture] }
    OtpVerifyRequest:
      type: object
      description: Accepts otp or the compatibility alias code; when both are supplied they must match (validated by the server). Real mode requires the server-issued challenge_id and its HttpOnly browser-binding cookie. Device fields are metadata only.
      required: [phone]
      anyOf:
        - required: [otp]
        - required: [code]
      properties:
        phone: { type: string, minLength: 6, maxLength: 32, pattern: '^\+?[0-9][0-9 -]{4,30}[0-9]$' }
        challenge_id: { type: string, format: uuid }
        otp: { type: string, minLength: 4, maxLength: 12 }
        code: { type: string, minLength: 4, maxLength: 12 }
        device_fingerprint: { type: string, minLength: 1, maxLength: 128 }
        device_id: { type: string, minLength: 1, maxLength: 128, description: Compatibility alias for device_fingerprint. }
    Session:
      type: object
      required: [id, device_id, expires_at]
      properties:
        id: { type: string, description: Public session reference; not a cookie credential and never usable to authenticate. }
        device_id: { type: string }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    LogoutRequest:
      type: object
      additionalProperties: false
      required: [operation_id]
      properties:
        operation_id: { type: string, format: uuid, description: One browser-generated operation ID reused for recovery of the same logout, never a login credential. }
    SessionResponse:
      type: object
      required: [user, session]
      properties:
        user:
          type: object
          required: [id]
          properties:
            id: { type: string }
            phone: { type: string, nullable: true }
        session: { $ref: '#/components/schemas/Session' }
    CurrentUserResponse:
      type: object
      required: [user_id, user, session]
      properties:
        user_id: { type: string }
        user:
          type: object
          required: [id]
          properties:
            id: { type: string }
            phone: { type: string, nullable: true }
        session: { $ref: '#/components/schemas/Session' }
    OtpVerifyResponse:
      type: object
      required: [user_id, user, session]
      properties:
        user_id: { type: string }
        user:
          type: object
          required: [id]
          properties:
            id: { type: string }
            phone: { type: string, nullable: true }
        session: { $ref: '#/components/schemas/Session' }
    SessionListResponse:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items: { $ref: '#/components/schemas/Session' }
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
    ErrorEnvelope:
      type: object
      required: [error_code, error_message]
      properties:
        error_code: { type: string, example: INGEST_TIMEOUT }
        error_message: { type: string }
        retriable: { type: boolean }
        details: { type: object, additionalProperties: true }
...
  responses:
    Error400:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error403:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error404:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error409:
      description: Conflict (Optimistic locking / idempotency)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error429:
      description: Too Many Requests (rate limited)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error503:
      description: Required authoritative store or provider unavailable; unknown external results remain unknown.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...